move StdGenerator to its own module
